+++ snocat/src/common/mod.rs
@@ -22,6 +22,7 @@ use std::{
 };
 
 pub mod authentication;
+pub mod daemon;
 pub mod protocol;
 pub mod tunnel_source;